Add a static_cast<byte>() around get_byte just to keep VC++ quiet.

+++ src/utils/loadstor.h
@@ -42,7 +42,9 @@ namespace Botan {
 */
 template<typename T> inline byte get_byte(u32bit byte_num, T input)
    {
-   return (input >> ((sizeof(T)-1-(byte_num&(sizeof(T)-1))) << 3));
+   return static_cast<byte>(
+      input >> ((sizeof(T)-1-(byte_num&(sizeof(T)-1))) << 3)
+      );
    }
 
 /*
